Add Windows compatibility and improve memory usage reporting
- Implemented compatibility for Windows in the codebase, including adjustments to format specifiers.
- Added a batch file for compiling algorithms with error handling.
- Updated memory usage reporting to use the correct format specifier for Windows.
- Included new executable files for various algorithms and added results files for performance analysis.
- Introduced a new source file for algorithm implementations and updated the main program to utilize these functions.

+++ busqueda_binaria/busqueda_binaria.c
@@ -129,7 +129,7 @@ void probar_busqueda_binaria() {
         
         // Mostrar resultados
         printf("  Tiempo de ejecución: %.6f ms\n", tiempo_ejecucion);
-        printf("  Memoria usada: %zu bytes\n", memoria_usada);
+        printf("  Memoria usada: %lu bytes\n", (unsigned long)memoria_usada);
         printf("  Elemento encontrado en índice: %d\n", resultado);
         
         // Liberar memoria
@@ -143,10 +143,10 @@ void probar_busqueda_binaria() {
     printf("\n=== RESUMEN BÚSQUEDA BINARIA ===\n");
     printf("Tamaño\tTiempo(ms)\tMemoria(bytes)\n");
     for (int i = 0; i < num_tamanos; i++) {
-        printf("%d\t%.6f\t%zu\n", 
+        printf("%d\t%.6f\t%lu\n", 
                mediciones[i].tamano_entrada,
                mediciones[i].tiempo_ejecucion,
-               mediciones[i].memoria_usada);
+               (unsigned long)mediciones[i].memoria_usada);
     }
     
     free(mediciones);
