+++ busqueda_lineal/busqueda_lineal.c
@@ -70,7 +70,7 @@ void probar_busqueda_lineal() {
         
         // Mostrar resultados
         printf("  Tiempo de ejecución: %.6f ms\n", tiempo_ejecucion);
-        printf("  Memoria usada: %zu bytes\n", memoria_usada);
+        printf("  Memoria usada: %lu bytes\n", (unsigned long)memoria_usada);
         printf("  Elemento encontrado en índice: %d\n", resultado);
         
         // Liberar memoria
@@ -84,10 +84,10 @@ void probar_busqueda_lineal() {
     printf("\n=== RESUMEN BÚSQUEDA LINEAL ===\n");
     printf("Tamaño\tTiempo(ms)\tMemoria(bytes)\n");
     for (int i = 0; i < num_tamanos; i++) {
-        printf("%d\t%.6f\t%zu\n", 
+        printf("%d\t%.6f\t%lu\n", 
                mediciones[i].tamano_entrada,
                mediciones[i].tiempo_ejecucion,
-               mediciones[i].memoria_usada);
+               (unsigned long)mediciones[i].memoria_usada);
     }
     
     free(mediciones);
